+++ packages/provider_kit/lib/provider_kit.dart
@@ -7,5 +7,4 @@ export 'src/errors/index.dart';
 export 'src/mutation/index.dart';
 export 'src/observer/index.dart';
 export 'src/state/index.dart';
-export 'src/utils/utils.dart';
 export 'src/view_state/index.dart';