+++ packages/provider_kit/lib/src/state/widgets/state_builder.dart
@@ -2,8 +2,8 @@ import 'package:flutter/foundation.dart';
 import 'package:flutter/widgets.dart';
 import 'package:provider/provider.dart';
 import 'package:provider_kit/src/base/state_value_listenable.dart';
+import 'package:provider_kit/src/state/type_defs/state_callbacks.dart';
 import 'package:provider_kit/src/state/widgets/state_listener.dart';
-import 'package:provider_kit/src/utils/type_definitions.dart';
 
 /// {@template provider_kit.stateBuilder}
 /// A widget that rebuilds its UI based on the state of a [StateValueListenable].