+++ packages/provider_kit/lib/src/state/widgets/state_listener.dart
@@ -3,8 +3,8 @@ import 'package:flutter/widgets.dart';
 import 'package:nested/nested.dart';
 import 'package:provider/provider.dart';
 import 'package:provider_kit/src/base/state_value_listenable.dart';
+import 'package:provider_kit/src/state/type_defs/state_callbacks.dart';
 import 'package:provider_kit/src/utils/equality_check.dart';
-import 'package:provider_kit/src/utils/type_definitions.dart';
 
 /// {@template provider_kit.stateListener}
 /// A widget that listens to a [StateValueListenable] and invokes a callback