+++ packages/provider_kit/lib/src/view_state/view_state_widgets.dart
@@ -1,8 +1,8 @@
 library provider_kit_view_state_widgets;
 
 import 'package:flutter/widgets.dart';
-import 'package:provider_kit/src/utils/type_definitions.dart';
 import 'package:provider_kit/src/view_state/notifiers/async_view_state_notifier.dart';
+import 'package:provider_kit/src/view_state/type_defs/view_state_callbacks.dart';
 import 'package:provider_kit/src/view_state/view_state_widgets_provider.dart';
 import 'package:provider_kit/src/view_state/states/view_states.dart';
 import 'package:flutter/foundation.dart';
@@ -10,7 +10,7 @@ import 'package:provider/provider.dart';
 import 'package:provider_kit/src/view_state/notifiers/view_state_notifier.dart';
 import 'package:provider_kit/src/state/index.dart';
 
-part 'view_state_base.dart';
+part 'view_state_widget_utils.dart';
 part '../view_state/widgets/view_state_builder.dart';
 part 'widgets/view_state_listener.dart';
 part 'widgets/view_state_consumer.dart';